test: opt out of new tests and refactor conflicting types
