+++ packages/client/tests/functional/_utils/setupTestSuiteMatrix.ts
@@ -1,13 +1,10 @@
-import { checkMissingProviders, MatrixOptions } from './checkMissingProviders'
+import { checkMissingProviders } from './checkMissingProviders'
 import { getTestSuiteConfigs, getTestSuiteMeta, getTestSuiteTable, TestSuiteConfig } from './getTestSuiteInfo'
 import { setupTestSuiteClient } from './setupTestSuiteClient'
 import { dropTestSuiteDatabase, setupTestSuiteDbURI } from './setupTestSuiteEnv'
+import { MatrixOptions } from './types'
 
 export type TestSuiteMeta = ReturnType<typeof getTestSuiteMeta>
-
-export type MatrixOptions = {
-  skipDb?: boolean
-}
 
 /**
  * How does this work from a high level? What steps?
