+++ packages/client/tests/functional/_utils/types.ts
@@ -1,0 +1,9 @@
+import { Providers } from './providers'
+
+export type MatrixOptions = {
+  optOut: {
+    from: `${Providers}`[]
+    reason: string
+  }
+  skipDb?: boolean
+}
