+++ packages/client/tests/functional/init-error/_matrix.ts
@@ -6,5 +6,13 @@ export default defineMatrix(() => [
       provider: 'postgresql',
       url: 'postgresql://invalid:invalid@invalid.org:123/invalid',
     },
+    {
+      provider: 'mysql',
+      url: 'mysql://invalid:invalid@invalid:3307/invalid',
+    },
+    {
+      provider: 'cockroachdb',
+      url: 'postgresql://invalid:invalid@invalid.org:123/invalid',
+    },
   ],
 ])
