+++ packages/client/tests/functional/init-error/tests.ts
@@ -5,6 +5,7 @@ import testMatrix from './_matrix'
 // @ts-ignore this is just for type checks
 declare let prisma: import('@prisma/client').PrismaClient
 
+// https://github.com/prisma/prisma/issues/10229
 testMatrix.setupTestSuite(
   () => {
     test('should assert that the error has the correct errorCode', async () => {
@@ -21,5 +22,14 @@ testMatrix.setupTestSuite(
       }
     })
   },
-  { skipDb: true }, // So we can maually call connect for this test
+  {
+    skipDb: true,
+    optOut: {
+      from: ['sqlite', 'mongodb'],
+      reason: `
+        sqlite dont have a connection string'
+        mongodb times out and dont throw
+      `,
+    },
+  },
 )
